requirements.txt
beautifulsoup4==4.12.3
darkdetect==0.8.0
deepgram_sdk==3.6.0
duckduckgo_search==6.3.3
google_api_python_client==2.146.0
googlesearch_python==1.2.5
groq==0.12.0
newsapi_python==0.2.7
nltk==3.9.1
openai==1.55.3
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.40.1
urllib3==2.2.3
websocket_client==1.8.0
